corrected server logic for login home screen. updated tsconfig to allow corresponding imports
